+++ tests/unit/tree/endpoint_node/test_validate.py
@@ -34,6 +34,7 @@ class TestValidate:
                 "delay",
                 tree_keys.VARS_KEY,
                 "options",
+                "proxies",
             ),
             ("name",),
             "endpoint",