+++ tests/unit/tree/request_node/test_validate.py
@@ -37,6 +37,7 @@ class TestValidate:
                 "delay",
                 "retry",
                 "options",
+                "proxies",
             ),
             ("name",),
             "request",