+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Implement new key `proxies` to request and endpoint nodes [#215](https://github.com/scanapi/scanapi/issues/215)
+
 ### Changed
 - Change requests to httpx. [#208](https://github.com/scanapi/scanapi/issues/208)